+++ src/main/java/techcourse/myblog/domain/ArticleRepository.java
@@ -7,23 +7,23 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.TreeMap;
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 @Repository
 public class ArticleRepository {
-    private static final int AUTO_INCREMENT_START_ID = 1;
-    private static final int MIN_ARTICLE_ID = 1;
+    private static final Long AUTO_INCREMENT_START_ID = 1L;
+    private static final Long MIN_ARTICLE_ID = 1L;
 
-    private AtomicInteger newArticleId = new AtomicInteger(AUTO_INCREMENT_START_ID);
+    private AtomicLong newArticleId = new AtomicLong(AUTO_INCREMENT_START_ID);
 
-    private Map<Integer, Article> articles = new TreeMap<>();
+    private Map<Long, Article> articles = new TreeMap<>();
 
     public List<Article> findAll() {
         return new ArrayList<>(articles.values());
     }
 
-    public int saveArticle(Article article) {
-        int id = article.getId();
+    public Long saveArticle(Article article) {
+        Long id = article.getId();
 
         if (isNewArticle(id)) {
             id = newArticleId.getAndIncrement();
@@ -34,25 +34,25 @@ public class ArticleRepository {
         return id;
     }
 
-    public Article findById(int articleId) {
+    public Article findById(Long articleId) {
         checkExistArticleId(articleId);
 
         return articles.get(articleId);
     }
 
-    public void removeArticle(int articleId) {
+    public void removeArticle(Long articleId) {
         checkExistArticleId(articleId);
 
         articles.remove(articleId);
     }
 
-    private void checkExistArticleId(int articleId) {
+    private void checkExistArticleId(Long articleId) {
         if (!articles.containsKey(articleId)) {
             throw new NotExistEntityException();
         }
     }
 
-    private boolean isNewArticle(int id) {
+    private boolean isNewArticle(Long id) {
         return id < MIN_ARTICLE_ID;
     }
 }
